+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * fix nested syntax rules with symbols after nested ellipsis
 * fix Dark Mode colors and scrolling of the page when using Bookmarklet on English Wikipedia
 * remove dependencies on `cond` from `-->` macro so you can use use `-->` inside `cond`
+* fix handling of recursive `flatten` `syntax-rules` macro [#304](https://github.com/jcubic/lips/issues/304)
 
 ## 1.0.0-beta.18
 ### Breaking